Added viewModel for change password view and refactored that view
