+++ MeuAtelieApp/MeuAtelieApp/MeuAtelieUI/Views/Profile/MAProfileView.swift
@@ -98,7 +98,7 @@ struct MAProfileView: View {
         }
         .ignoresSafeArea()
         .sheet(isPresented: $viewModel.isShowingChangePassword) {
-            MANewPassword()
+            MANewPasswordView()
         }
     }
